21-06-2023

+++ 057.py
@@ -1,7 +1,7 @@
 peso = []
 
 for c in range(0,5):
-    peso.append(int(input('Peso: ')))
+    peso.append(float(input('Peso: ')))
 
 print(max(peso))
 print(min(peso))
